[Feature #128843401] fraction team member into a separate file
