+++ codango/static/js/components/team-new.jsx
@@ -1,28 +1,7 @@
 import React, { Component } from 'react';
 import { Grid, Row, Tabs, Tab } from 'react-bootstrap';
+import TeamMember from './team-member.jsx';
 var Slider = require('react-slick');
-
-
-class TeamMember extends Component {
-  render() {
-    return (
-      <div>
-        <div className="team-img-container">
-          <img src={this.props.imgSrc} />
-          <div className="team-social-container">
-            <div className="team-social-wrapper">
-              <a href={this.props.twitter}><span className="mdi mdi-twitter"></span></a>
-              <a href={this.props.github}><span className="mdi mdi-github-box"></span></a>
-              <a href={this.props.linkedin}><span className="mdi mdi-linkedin"></span></a>
-            </div>
-          </div>
-        </div>
-        <a href={this.props.github}><h3>{this.props.name}</h3></a>
-        <p>{this.props.position}</p>
-      </div>
-    )
-  }
-}
 
 
 class Team extends Component {
